+++ src/constants/additional-audio-video-devices.ts
@@ -1,0 +1,10 @@
+export const VIDEO_INPUT = {
+  NONE: 'None',
+  BLUE: 'Blue',
+  SMPTE: 'SMPTE Color Bars',
+};
+
+export const AUDIO_INPUT = {
+  NONE: 'None',
+  440: '440 Hz',
+};